Makes analyse a method of document and moves code around to match this behaviour.
