+++ searcher/resource/resource.go
@@ -1,5 +1,11 @@
 package resource
 
+import (
+	"log"
+
+	"github.com/sleepypioneer/resource-searcher/searcher/analyse"
+)
+
 type Resource struct {
 	Document *Document `json:"document"`
 	Meta     *Meta     `json:"meta"`
@@ -20,4 +26,14 @@ type Document struct {
 	Content   []string `json:"content"`
 	WordCount int
 	Topics    []string
+}
+
+// Analyse runs the various analysing logic on the give document
+func (doc *Document) Analyse() error {
+	// do something with your document
+	doc.WordCount = analyse.WordCount(doc.Content)
+	doc.Topics = analyse.FindTopics(doc.Content)
+	log.Printf("doc word count: %d", doc.WordCount)
+	log.Printf("doc topics: %v", doc.Topics)
+	return nil
 }
